Egen config for session repo
