+++ apps/dolly-frontend/src/main/java/no/nav/dolly/web/config/ProdConfig.java
@@ -1,5 +1,6 @@
 package no.nav.dolly.web.config;
 
+import no.nav.testnav.libs.reactivesessionsecurity.config.OidcSessionRepositoryConfiguration;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 import org.springframework.context.annotation.Profile;
@@ -9,7 +10,8 @@ import no.nav.testnav.libs.reactivesessionsecurity.config.OicdRedisSessionConfig
 @Configuration
 @Profile("prod")
 @Import({
-        OicdRedisSessionConfiguration.class
+        OicdRedisSessionConfiguration.class,
+        OidcSessionRepositoryConfiguration.class
 })
 public class ProdConfig {
 }
