+++ libs/reactive-session-security/src/main/java/no/nav/testnav/libs/reactivesessionsecurity/config/OicdInMemorySessionConfiguration.java
@@ -1,14 +1,9 @@
 package no.nav.testnav.libs.reactivesessionsecurity.config;
 
-import no.nav.testnav.libs.reactivesessionsecurity.repository.OidcReactiveMapSessionRepository;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.boot.autoconfigure.session.SessionProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Import;
 import org.springframework.session.config.annotation.web.server.EnableSpringWebSession;
-
-import java.time.Duration;
-import java.util.concurrent.ConcurrentHashMap;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -31,20 +26,6 @@ import no.nav.testnav.libs.reactivesessionsecurity.resolver.InMemoryTokenResolve
 })
 @RequiredArgsConstructor
 public class OicdInMemorySessionConfiguration {
-
-    private final SessionProperties sessionProperties;
-
-    @Bean
-    public OidcReactiveMapSessionRepository reactiveSessionRepository() {
-        OidcReactiveMapSessionRepository sessionRepository = new OidcReactiveMapSessionRepository(new ConcurrentHashMap<>());
-        int defaultMaxInactiveInterval = (int) (sessionProperties.getTimeout() == null
-                ? Duration.ofMinutes(30)
-                : sessionProperties.getTimeout()
-        ).toSeconds();
-        sessionRepository.setDefaultMaxInactiveInterval(defaultMaxInactiveInterval);
-        log.info("Set in-memory session max inactive to {} seconds.", defaultMaxInactiveInterval);
-        return sessionRepository;
-    }
 
     @Bean
     @ConditionalOnMissingBean
